pyproject.toml
[tool.poetry]
name = "prompt-author"
version = "0.1.0"
description = "AI powered author that helps you write various types of content."
authors = ["Gijs Wobben <gijswobben@gmail.com>"]
readme = "docs/README.md"
packages = [{ include = "prompt_author" }]

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.6"
langchain = "0.0.277"
openai = "^0.27.8"
tiktoken = "^0.4.0"
pydantic = "^2.3.0"
rich = "^13.5.2"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
mypy = "^1.4.1"
ruff = "^0.0.284"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
prompt-author = "prompt_author.cli:main"

[tool.pytest.ini_options]
addopts = "--cov=prompt_author --cov-report term-missing"

[tool.coverage.run]
omit = ["prompt_author/cli.py"]
